refactor: remove GamesUI component and integrate its functionality into Games component
