+++ src/App.jsx
@@ -15,7 +15,6 @@ import PartyChatUI from "./components/Games/PartyChatUI";
 import Price from "./components/Price";
 import Profile from "./components/Profile/Profile";
 import PartiesUI from "./components/Games/PartiesUI";
-import GamesUI from "./components/Games/GamesUI";
 
 function App() {
     return (
@@ -30,7 +29,6 @@ function App() {
                     <Route path="partychat" element={<PartyChatUI />} />
                     <Route element={<ProtectedRoute />}>
                         <Route path="games" element={<Games />} />
-                        <Route path="gameso" element={<GamesUI />} />
                         <Route path="games/:id" element={<Parties />} />
                         <Route path="gameso/:id" element={<PartiesUI />} />
                         <Route
